Rename GenerateAnalyzerRulesets.csproj to GenerateDocumentationAndConfigFiles.csproj
Over time this post-build tool has been enhanced from just a ruleset generator to generating bunch of documentation files (md + sarif) and configuration files (rulesets + editorconfigs). About time this utility got renamed appropriately.
